refactor(project-cut): converge core imports on stable entrypoint
