+++ framework/assignment-capture/assignment-capture.test.mjs
@@ -8,7 +8,7 @@ import os from 'node:os';
 import path from 'node:path';
 import { test } from 'node:test';
 import { fileURLToPath } from 'node:url';
-import { semanticRoot } from '../project-cut/src/project-cut.mjs';
+import { semanticRoot } from '../project-cut/index.mjs';
 import {
   ASSIGNMENT_REQUEST_SCHEMA,
   AssignmentCaptureError,
