+++ framework/maintainability/terminal-evidence.mjs
@@ -9,7 +9,7 @@ import path from 'node:path';
 import process from 'node:process';
 import { fileURLToPath } from 'node:url';
 
-import { semanticRoot } from '../project-cut/src/project-cut.mjs';
+import { semanticRoot } from '../project-cut/index.mjs';
 import { collectTerminalLiveObservations } from '../terminal-evidence/live-observations.mjs';
 import {
   verifyLiveAssignment,
