+++ framework/terminal-evidence/live-observations.mjs
@@ -7,7 +7,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import { semanticRoot } from '../project-cut/src/project-cut.mjs';
+import { semanticRoot } from '../project-cut/index.mjs';
 
 const ROOT = path.resolve(
   path.dirname(fileURLToPath(import.meta.url)),
