+++ framework/work-design-advisor/src/work-design-advisor.mjs
@@ -1,10 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // @ts-check
 
-import {
-  canonicalJson,
-  semanticRoot,
-} from '../../project-cut/src/project-cut.mjs';
+import { canonicalJson, semanticRoot } from '../../project-cut/index.mjs';
 
 export const WORK_DESIGN_REQUEST_SCHEMA =
   'kungfu.work-design.advice-request/v1';
