+++ framework/work-design-preflight/tooling/check-work-design-preflight.mjs
@@ -5,7 +5,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import { semanticRoot } from '../../project-cut/src/project-cut.mjs';
+import { semanticRoot } from '../../project-cut/index.mjs';
 import {
   workDesignAuthorityBoundary,
   workDesignAutoAdoptionPolicy,
