+++ framework/work-design-preflight/work-design-runtime.json
@@ -3,6 +3,7 @@
   "entrypoint": "work-design-preflight/tooling/work-design-preflight.mjs",
   "files": [
     "format/project-cut-canonical-json.mjs",
+    "project-cut/index.mjs",
     "project-cut/src/json-scanner.mjs",
     "project-cut/src/project-cut.mjs",
     "work-design-advisor/src/work-design-advisor.mjs",
