+++ framework/work-history-selector/tooling/work-history-selector-contract.mjs
@@ -6,7 +6,7 @@ import { createRequire } from 'node:module';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import { semanticRoot } from '../../project-cut/src/project-cut.mjs';
+import { semanticRoot } from '../../project-cut/index.mjs';
 
 const DEFAULT_ROOT = path.resolve(
   path.dirname(fileURLToPath(import.meta.url)),
