+++ scripts/backfill-legacy-atlas-roots.mjs
@@ -16,7 +16,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath, pathToFileURL } from 'node:url';
 
-import { canonicalJson } from '../framework/project-cut/src/project-cut.mjs';
+import { canonicalJson } from '../framework/project-cut/index.mjs';
 
 const ROOT = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
 const OUTPUT = path.join(
