+++ scripts/backfill-legacy-atlas-roots.test.mjs
@@ -7,7 +7,7 @@ import os from 'node:os';
 import path from 'node:path';
 import test from 'node:test';
 
-import { canonicalJson } from '../framework/project-cut/src/project-cut.mjs';
+import { canonicalJson } from '../framework/project-cut/index.mjs';
 import { legacyAtlasRoots } from './backfill-legacy-atlas-roots.mjs';
 
 function sha(value) {
