+++ scripts/buildchain-documentation-witness.test.mjs
@@ -8,7 +8,7 @@ import path from 'node:path';
 import test from 'node:test';
 import { fileURLToPath } from 'node:url';
 
-import { canonicalJson } from '../framework/project-cut/src/project-cut.mjs';
+import { canonicalJson } from '../framework/project-cut/index.mjs';
 import {
   LEGACY_ATLAS_ROOTS_DIGEST,
   LEGACY_ATLAS_ROOTS_PATH,
