+++ scripts/check-project-cut-settlement.test.mjs
@@ -16,7 +16,7 @@ import {
 import {
   canonicalJson,
   semanticRoot,
-} from '../framework/project-cut/src/project-cut.mjs';
+} from '../framework/project-cut/index.mjs';
 import {
   abandonSettlement,
   inspectSettlement,
